+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Improvements
 
 - DAPI-19: Add filter on subscribed/unsubscribed products to Franklin Insights in the product grid
+- DAPI-20: Display Franklin Insights subscription as a column of the product grid
 
 ## Bug fixes